refactor dashboard app
