--- euth/a4_dashboard/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class A4DashboardConfig(AppConfig):
-    name = 'euth.a4_dashboard'
-    label = 'euth_a4_dashboard'
