+++ euth/contrib/static/js/app.js
@@ -4,7 +4,6 @@ var ReactParagraphs = require('../../../documents/static/documents/ParagraphBox.
 var ReactRatings = require('adhocracy4').ratings
 var ReactFollow = require('../../../follows/static/follows/react_follows.jsx')
 var ReactLanguageSwitch = require('../../../dashboard/static/language_switch/react_language_switch.jsx')
-var ReactUserList = require('../../../dashboard/static/user_list/react_user_list.jsx')
 var $ = window.jQuery = window.$ = require('jquery')
 
 require('../../../../euth_wagtail/static/js/euth_wagtail')
@@ -14,6 +13,5 @@ $(function () {
   widget.initialise('a4', 'ratings', ReactRatings.renderRatings)
   widget.initialise('euth', 'document', ReactParagraphs.renderParagraphs)
   widget.initialise('euth', 'follows', ReactFollow.renderFollow)
-  widget.initialise('euth', 'userlist', ReactUserList.renderUserList)
   widget.initialise('euth', 'language-switch', ReactLanguageSwitch.renderLanguageSwitch)
 })
