--- euth/dashboard/templates/euth_dashboard/base_organisation.html
@@ -1,53 +1,0 @@
-{% extends "base.html" %}
-{% load i18n %}
-
-{% block meta_viewport %}{% endblock %}
-
-{% block content %}
-<div class="container">
-    <div class="dashboard row">
-        <div class="col-md-3">
-            <div class="dashboard-nav">
-                <div class="dashboard-nav-group">
-                    {% include 'euth_dashboard/includes/dropdown.html' with logo=view.organisation.logo title=view.organisation.name orgs=view.other_organisations_of_user%}
-                    <h4 class="h6 sans-serif text-uppercase dashboard-nav-title">{% trans 'Organisation settings' %}</h4>
-
-                    <!-- Organisation preferences -->
-                    <div class="list-group">
-                        {% url 'dashboard-organisation-edit' organisation_slug=view.organisation.slug as dashboard_organisations_edit %}
-                        <a href="{{dashboard_organisations_edit}}"
-                           class="list-group-item list-group-item-heading {% if view.menu_item == 'organisation' %}selected{% endif %}">
-                            <i class="fa fa-rocket" aria-hidden="true"></i> <span>{% trans 'Organisation information' %}</span>
-                        </a>
-
-                        {% url 'dashboard-project-list' organisation_slug=view.organisation.slug as dashboard_project_list %}
-                        {% url 'dashboard-blueprint-list' organisation_slug=view.organisation.slug as dashboard_blueprint_list %}
-                        <a href="{{dashboard_project_list}}"
-                           class="list-group-item list-group-item-heading {% if view.menu_item == 'project' %}selected{% endif %}">
-                            <i class="fa fa-clone" aria-hidden="true"></i> <span>{% trans 'Projects' %}</span>
-                        </a>
-
-                        <!--
-                        <a href="#" class="list-group-item list-group-item-heading">
-                            <i class="fa fa-check" aria-hidden="true"></i> <span>Moderation</span>
-                        </a>
-                        <a href="#" class="list-group-item">
-                            <h4 class="list-group-item-heading">
-                                <i class="fa fa-users" aria-hidden="true"></i> Users
-                            </h4>
-                        </a>-->
-
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="col-xs-9">
-            <div class="dashboard-content">
-                {% block dashboard_content %}{% endblock %}
-            </div>
-        </div>
-    </div>
-</div>
-
-
-{% endblock %}
