--- euth/dashboard/templates/euth_dashboard/blueprint_list.html
@@ -1,68 +1,0 @@
-{% extends "euth_dashboard/base_organisation.html" %}
-{% load widget_tweaks i18n static %}
-
-
-{% block dashboard_content %}
-    <h1 class="black-divider dashboard-title">{% trans 'Create new project from one of the templates' %}</h1>
-    <div class="action-bar">
-        <nav class="navbar navbar-default navbar-static">
-            <ul class="nav navbar-nav">
-                <li>
-                    <a href="{% url 'dashboard-project-list' view.organisation.slug %}">
-                        <i class="fa fa-arrow-left" aria-hidden="true"></i>
-                        {% trans 'back to project list' %}</a>
-                </li>
-            </ul>
-        </nav>
-    </div>
-
-    <div class="blueprint-list">
-
-        <div class="dst-tile">
-            <img src="{% static 'images/placeholder_transparent.png' %}" alt="Decision support tool">
-            <h2 class="sans-serif h4">{% trans 'Not sure which template?' %}</h2>
-            <div class="dst-description">
-                <h3 class="h5">{% trans "Try the Decision Support Tool (DST)!" %}</h3>
-                {% trans "In just three minutes you'll find out which process suits your needs. It's easy." %}
-            </div>
-
-            <div class="dst-footer">
-                <a href="{% url 'blueprints-form' organisation_slug=view.organisation.slug %}"
-                   class="btn btn-primary">
-                    {% trans 'Choose template'%}
-                </a>
-            </div>
-        </div>
-
-        {% for blueprint_slug, blueprint in view.blueprints %}
-
-        <div class="blueprint-tile">
-            <img class="blueprint-image" src="{% static blueprint.image %}" alt="{{ blueprint.title }}">
-            <h2 class="sans-serif h4">{{ blueprint.title }}</h2>
-            <div class="blueprint-description">
-                {{ blueprint.description | linebreaks }}
-            </div>
-
-            <div class="blueprint-phases">
-                <ul class="list-unstyled">
-                    {% for phase_content in blueprint.content %}
-                    <li><strong>Phase {{ forloop.counter }}</strong>: {{ phase_content.description }}</li>
-                    {% endfor %}
-                </ul>
-            </div>
-            <div class="blueprint-footer">
-                <a href="{% url 'dashboard-project-create' organisation_slug=view.organisation.slug blueprint_slug=blueprint_slug %}"
-                   class="btn btn-primary">
-                    {% trans 'use this template' %}
-                </a>
-
-                <button class="btn-link" type="button" data-toggle="modal" data-target="#dst-infopage-{{ blueprint_slug }}">{% trans 'Learn More' %}</button>
-            </div>
-
-            {% include 'euth_blueprints/includes/infopage.html'%}
-
-        </div>
-        {% endfor %}
-</div>
-
-{% endblock %}
