--- euth/dashboard/templates/euth_dashboard/includes/dropdown.html
@@ -1,29 +1,0 @@
-{% load thumbnail avatar %}
-
-<div class="dropdown">
-    <button id="dLabel" type="button" data-toggle="dropdown" aria-haspopup="true"
-            aria-expanded="false">
-        {% if logo %}
-            <span class="circle-bg">
-                <img src="{{ logo|thumbnail_url:'avatar_small' }}" alt="">
-            </span>
-        {% elif is_user %}
-            <span class="circled"
-                  style="background-image: url({% get_avatar request.user 'avatar_small' %});"></span>
-        {% endif %}
-        <span class="dropdown-title">{{ title }}</span>
-        {% if orgs %}
-            <span class="caret"></span>
-        {% endif %}
-    </button>
-    {% if orgs %}
-        <ul class="dropdown-menu" aria-labelledby="dLabel">
-            {% for organisation in orgs %}
-                <li>
-                    <a href="{% url 'dashboard-project-list' organisation_slug=organisation.slug %}">{{ organisation.name }}</a>
-                </li>
-            {% endfor %}
-        </ul>
-    {% endif %}
-</div>
-<hr>
