--- euth/dashboard/templates/euth_dashboard/includes/inline_form.html
@@ -1,28 +1,0 @@
-{% load i18n widget_tweaks %}
-<div class="form-group category-form {% if form.name.errors %} has-error{% endif %}"
-     data-container="body"
-     data-toggle="popover"
-     data-trigger="hover"
-     data-placement="top"
-     data-content="{% for error in form.name.errors %}{{ error }}{% endfor %}">
-
-    {% if can_delete %}
-        <input type="checkbox" name="{{ form.DELETE.html_name }}" id="{{ form.DELETE.id_for_label }}" class="category__delete">
-    {% endif %}
-
-    <label for="{{ form.name.id_for_label }}" class="sr-only">
-        {{ form.name.label }}
-    </label>
-
-    {% if can_delete %}
-        <label for="{{ form.DELETE.id_for_label }}" class="btn btn-danger category__delete-label" title="{% trans 'Remove category' %}">
-            <i class="fa fa-times" aria-label="{% trans 'Remove category' %}"></i>
-        </label>
-    {% endif %}
-
-    {% render_field form.name class="form-control category-text" %}
-
-    {% for hidden in form.hidden_fields %}
-        {{ hidden }}
-    {% endfor %}
-</div>
