--- euth/dashboard/templates/euth_dashboard/includes/project_form_information_tab.html
@@ -1,41 +1,0 @@
-{% load i18n form_tags widget_tweaks %}
-<article id="information" class="tab-panel {% if not view.project.active_phase %} active {% endif %}">
-    {% if field.errors %}
-    <ul class="errorlist">
-    {% for error in form.project.non_field_errors %}
-      <li> {{ error }} </li>
-    {% endfor %}
-    </ul>
-    {% endif %}
-
-    {% for name, fields in form.project.formsections.information.items %}
-    <div class="form-section">
-        <h3 class="form-section-heading">{{ name }}</h3>
-
-        <div class="form-section-fields">
-        {% for field in form.project %}
-        {% if field.name in fields %}
-        <div class="form-group{% if field.errors %} has-error{% endif %}">
-            <label for="{{ field.id_for_label }}">
-                {{ field.label }} {% if field.field.required %}*{% endif %}
-            </label>
-            <p class='form-help-text'>{{ field.help_text }}</p>
-            {% if field|widget_type == 'radioselect' %}
-                {% render_field field %}
-            {% else %}
-                {% render_field field class="form-control" %}
-            {% endif %}
-            {% if field.errors %}
-            <ul class="errorlist">
-            {% for error in field.errors %}
-              <li> {{ error }} </li>
-            {% endfor %}
-            </ul>
-            {% endif %}
-        </div>
-        {% endif %}
-        {% endfor %}
-        </div>
-    </div>
-    {% endfor %}
-</article>
