--- euth/dashboard/templates/euth_dashboard/includes/project_form_result_tab.html
@@ -1,18 +1,0 @@
-{% load i18n widget_tweaks %}
-
-<article id="result" class="tab-panel">
-
-    <div class="form-section">
-        <h3 class="form-section-heading">{% trans 'Result'%}</h3>
-        <div class="form-section-fields">
-
-            <div class="form-group{% if field.errors %} has-error{% endif %}">
-                <label for="{{form.project.result.id_for_label}}">{{ form.project.result.label }}</label>
-                <p class='form-help-text'>{{ form.project.result.help_text }}</p>
-                {% render_field form.project.result class="form-control" %}
-                <span class="help-block">{{ error }}</span>
-            </div>
-        </div>
-    </div>
-
-</article>
