--- euth/dashboard/templates/euth_dashboard/includes/project_form_tabs.html
@@ -1,17 +1,0 @@
-{% load i18n form_tags %}
-<nav class="tabs">
-    <ul>
-        <li class="col-md-4 active">
-            <a href="#information" data-toggle="tab" class="tab">{% trans "Information" %}</a>
-            {% if form.project.errors %}<span class="info-number">{{ form.project.errors|length }}</span>{% endif %}
-        </li>
-        <li class="col-md-4">
-            <a href="#participation" data-toggle="tab" class="tab">{% trans "Participation" %}</a>
-            {% add form.phases.total_error_count form.module_settings.errors|length as error_count %}
-            {% if error_count > 0  %}<span class="info-number">{{ error_count }}</span>{% endif %}
-        </li>
-        <li class="col-md-4"><a href="#result" data-toggle="tab" class="tab">{% trans "Result" %}</a>
-            {% if form.project.result.errors %}<span class="info-number">{{ from.project.result.errors|length }}</span>{% endif %}
-        </li>
-    </ul>
-</nav>
