--- euth/dashboard/templates/euth_dashboard/organisation_form.html
@@ -1,52 +1,0 @@
-{% extends "euth_dashboard/base_organisation.html" %}
-{% load widget_tweaks i18n %}
-
-{% block dashboard_content %}
-<h1 class="black-divider dashboard-title">{% trans 'Edit your organisation information' %}</h1>
-
-<div class="general-form">
-    <form enctype="multipart/form-data" action="{{ request.path }}" method="post">
-        {% csrf_token %}
-
-        {{ form.media }}
-        {{ form.non_field_errors }}
-
-        {% for field in form.untranslated %}
-        <div class="form-group{% if field.errors %} has-error{% endif %}">
-            <label for="{{field.id_for_label}}">{{ field.label }}</label>
-            {% if field.help_text %}
-            <p><i>{{ field.help_text }}</i></p>
-            {% endif %}
-            {% render_field field class="form-control" %}
-            {% for error in field.errors %}
-            <span class="help-block">{{ error }}</span>
-            {% endfor %}
-        </div>
-        {% endfor %}
-        <div class="language-switch-list"
-             id="language-switch-list"
-             data-euth-widget="language-switch"
-             data-languages="{{ form.languages|join:' ' }}"
-             data-active-languages="{{ form.prefiled_languages|join:' ' }}"
-        ></div>
-
-        {% for lang_code, fields in form.translated %}
-        <div class="language-switch-panel {% if forloop.first %}active{% endif %}" id="{{ lang_code }}_language_panel">
-            {% for field in fields %}
-            <div class="form-group{% if field.errors %} has-error{% endif %}">
-                <label for="{{field.id_for_label}}">{{ field.label }}</label>
-                {% if field.help_text %}
-                <p><i>{{ field.help_text }}</i></p>
-                {% endif %}
-                {% render_field field class="form-control" %}
-                {% for error in field.errors %}
-                <span class="help-block">{{ error }}</span>
-                {% endfor %}
-            </div>
-            {% endfor %}
-        </div>
-        {% endfor %}
-
-        <button type="submit" name="publish" class="btn btn-sm btn-primary">{% trans 'save changes' %}</button>
-
-{% endblock %}
