--- euth/dashboard/templates/euth_dashboard/project_form.html
@@ -1,99 +1,0 @@
-{% extends "euth_dashboard/base_organisation.html" %}
-{% load static widget_tweaks i18n form_tags flashpoll_form_tags %}
-{% block dashboard_content %}
-{% get_disabled project as disabled %}
-    <h1 class="black-divider dashboard-title dashboard-title-no-margin">{{ heading }} <em>{% if view.blueprint %}{{ view.blueprint.title }}{% else %}{{ project.name }}{% endif %}</em></h1>
-
-<div class="general-form dashboard-tabbed-form">
-    <form enctype="multipart/form-data" action="{{ request.path }}" method="post">
-
-        {% if form.is_bound and not form.is_valid %}
-        <div class="errorbar">
-            {% trans 'Unable to submit form. Please fix the errors.' %}
-        </div>
-        {% elif project and project.is_archived %}
-        <div class="infobar">
-            {% trans 'This project is archived.' %}
-        </div>
-        {% endif %}
-
-        {% csrf_token %}
-
-        {% include 'euth_dashboard/includes/project_form_tabs.html' with form=form view=view %}
-
-        {% include 'euth_dashboard/includes/project_form_information_tab.html' with form=form view=view %}
-
-        {% if module_settings == "euth_flashpoll" %}
-            {% include 'euth_flashpoll/includes/project_form_participation_flashpoll_tab.html' with form=form view=view %}
-            {% include 'euth_flashpoll/includes/project_form_result_flashpoll_tab.html' with form=form view=view %}
-        {% else %}
-            {% include 'euth_dashboard/includes/project_form_participation_tab.html' with form=form view=view %}
-            {% include 'euth_dashboard/includes/project_form_result_tab.html' with form=form view=view %}
-        {% endif %}
-
-        <div class="dashboard-form-buttons dropdown">
-            <div class="btn-group">
-
-                {% if not project or project.is_draft %}
-                <button type="submit" name="save_draft" class="btn btn-primary" {{ disabled }}>
-                    <span class="dropdown-text">{% trans 'save draft' %}</span>
-                </button>
-                {% else %}
-                <button type="submit" name="publish" class="btn btn-primary" {{ disabled }}>
-                    <span class="dropdown-text">{% trans 'save and publish' %}</span>
-                </button>
-                {% endif %}
-
-                <button class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                    <span class="caret"></span>
-                </button>
-
-                <ul class="dropdown-menu">
-                    <li>
-                        {% if not project or project.is_draft %}
-                        <button type="submit" name="publish" class="btn btn-sm btn-hover-primary" {{ disabled }}>
-                            <i class="fa fa-fw fa-flag" aria-hidden="true"></i>
-                            <span class="dropdown-text">{% trans 'save and publish' %}</span>
-                        </button>
-                        {% else %}
-                        <button type="submit" name="save_draft" class="btn btn-sm btn-hover-primary" {{ disabled }}>
-                            <i class="fa fa-fw fa-pause"> </i>
-                            <span class="dropdown-text">{% trans 'unpublish' %}</span>
-                        </button>
-                        {% endif %}
-                    </li>
-                    {% if module_settings == "euth_flashpoll" %}
-                    {% with poll=form.module_settings.poll.value|get_poll_from_string %}
-                    {% if project.pk != None %}
-                    <li>
-                        <button type="button" class="btn btn-sm btn-hover-danger" data-toggle="modal" data-target="#deleteModal">
-                            <i class="fa fa-fw fa-trash-o" aria-hidden="true"></i>
-                            <span class="dropdown-text">{% trans 'delete' %}</span>
-                        </button>
-                    </li>
-                    {% endif %}
-                    {% endwith %}
-                    {% elif project.pk != None %}
-                    <li><button type="button" class="btn btn-sm btn-danger btn-hover-danger"  data-toggle="modal" data-target="#deleteModal"><i class="fa fa-fw fa-trash-o" aria-hidden="true"></i><span class="dropdown-text">{% trans 'delete' %}</span></button></li>
-                    {% endif %}
-                </ul>
-            </div>
-        </div>
-    </form>
-</div>
-
-{% url 'dashboard-project-delete' project.slug as delete_view_url %}
-{% include 'includes/delete_modal.html' with name=project.name delete_action=delete_view_url %}
-
-{% endblock %}
-
-{% block extra_css %}
-{{ block.super }}
-{{ form.media.css }}
-{% endblock %}
-
-{% block extra_js %}
-{{ block.super }}
-{{ form.media.js }}
-<script type="text/javascript" src="{% static 'add_offline_phase/add_offline_phase.js' %}"></script>
-{% endblock %}
