--- euth/dashboard/templates/euth_dashboard/project_invites.html
@@ -1,18 +1,0 @@
-{% extends "euth_dashboard/base_organisation.html" %}
-{% load widget_tweaks i18n %}
-
-{% block dashboard_content %}
-<h1 class="black-divider dashboard-title">{% trans 'Invite users to project' %} <em>{{ project }}</em></h1>
-<form method="POST" class="general-form">
-    {% csrf_token %}
-    <div class="form-group">
-        <label for="{{ form.emails.id_for_label }}">{{ form.emails.label }}</label>
-        <p class='form-help-text'>{{ form.emails.help_text }}</p>
-        {{ form.emails|add_class:'form-control' }}
-        {{ form.emails.errors }}
-    </div>
-    <input type="submit" class="btn btn-primary" value="{% trans 'send invitations' %}">
-</form>
-{% endblock %}
-
-{% block extra_js %}{{ form.media }}{% endblock %}
