--- euth/dashboard/templates/euth_dashboard/project_list.html
@@ -1,99 +1,0 @@
-{% extends "euth_dashboard/base_organisation.html" %}
-{% load project_tags idea_tags widget_tweaks i18n %}
-
-{% block dashboard_content %}
-<h1 class="black-divider dashboard-title">{% trans 'Projects' %}</h1>
-
-<div class="dashboard-project-controls">
-    <a href="{% url 'dashboard-blueprint-list' organisation_slug=view.organisation.slug %}" class="dashboard-project-create btn btn-primary">{% trans 'create project' %}</a>
-    <div class="dashboard-project-filter filter-bar">
-        {% for field in view.filter.form %}
-        {% if field.name != 'ordering' %}
-        {{ field }}
-        {% endif %}
-        {% endfor %}
-    </div>
-</div>
-
-<div class="dashboard-projects-list">
-    {% if project_list|length == 0 %}
-    <div class="infotext">
-      {% trans 'No projects found' %}
-    </div>
-
-    {% else %}
-
-    <ul>
-        {% for project in project_list %}
-            <li class="tile-list project-tile-list">
-                <div class="row">
-                    <div class="col-sm-8">
-                        <a href="{% url 'dashboard-project-edit' project.slug %}">
-                        <i class="fa fa-edit" aria-hidden="true"></i>
-                        <strong>
-                            {{ project.name }}
-                        </strong>
-                    </a>
-                        {% if project.active_phase %}
-                            <span class="text-muted">{% trans 'Phase' %} {{ project.module_set.first.phases_passed|length|add:1 }}: {{ project.active_phase.content.name }}</span>
-                        {% endif %}
-
-                        {% get_days project.days_left as days %}
-                        <div class="project-tile-badges">
-                            <span class="label label-type">{{ project.module_set.first.phase_set.first.content.module_name }}</span>
-                            {% if not project.is_public %}
-                            <span class="label label-private">{% trans 'private' %}</span>
-                            {% endif %}
-                            {% if project.is_archived %}
-                            <span class="label label-archived">{% trans 'archived' %}</span>
-                            {% elif project.has_finished %}
-                            <span class="label label-finished">{% trans 'finished' %}</span>
-                            {% endif %}
-                            {% if days %}
-                            <span class="label label-runningout">{{ days }}</span>
-                            {% endif %}
-                        </div>
-                    </div>
-
-                    <div class="col-sm-4">
-                        <div class="pull-right">
-                            <i class="fa fa-star"></i> {{ project.follow_set.all.count }}
-                            <a href="{% url 'dashboard-project-users' project.slug %}"
-                               title="{% trans 'Users' %}" class="btn btn-hover-primary btn-sm">
-                                    <i class="fa fa-user" aria-hidden="true"></i>
-                            </a>
-                            <a href="{% url 'project-detail' project.slug %}" class="btn btn-hover-primary btn-sm" title="{% trans 'view project' %}">
-                                <i class="fa fa-eye" aria-hidden="true"></i>
-                            </a>
-                            {% for module in project.module_set.all %}
-                            {% is_idea_list module as has_ideas  %}
-                            {% if has_ideas %}
-                            <a href={% url 'idea-download' module.slug %} class="btn btn-hover-primary btn-sm" title="{% trans 'download module' %}">
-                                <i class="fa fa-download" aria-hidden="true"></i>
-                            </a>
-                            {% endif %}
-                            {% endfor %}
-                            {% if project.is_archivable %}
-                            <form action="{% url 'dashboard-project-archive' project.slug %}" method="post">
-                              {% csrf_token %}
-                              <button type="submit" class="btn btn-sm btn-hover-primary" title="{% trans 'archive project' %}">
-                                <i class="fa fa-archive" aria-hidden="true"></i>
-                              </button>
-                            </form>
-                            {% elif project.is_archived %}
-                            <form action="{% url 'dashboard-project-unarchive' project.slug %}" method="post">
-                              {% csrf_token %}
-                              <button type="submit" class="btn btn-sm btn-hover-primary" title="{% trans 'unarchive project' %}">
-                                <i class="fa fa-archive" aria-hidden="true"></i>
-                              </button>
-                            </form>
-                            {% endif %}
-                        </div>
-                    </div>
-                </div>
-            </li>
-        {% endfor %}
-    </ul>
-</div>
-{% endif %}
-{% endblock %}
