--- euth/dashboard/templates/euth_dashboard/project_users.html
@@ -1,137 +1,0 @@
-{% extends "euth_dashboard/base_organisation.html" %}
-{% load widget_tweaks i18n react_user_list avatar %}
-
-{% block dashboard_content %}
-
-    <h1 class="black-divider dashboard-title">
-        {% if project.is_private %}
-            {% trans 'Manage users of' %}
-        {% else %}
-            {% trans 'Manage moderators of' %}
-        {% endif %}
-        <em>{{ project.name }}</em>
-    </h1>
-
-<form method="POST">
-{% csrf_token %}
-{% if not project.is_public %}
-    <ul class="nav nav-tabs" role="tablist">
-        <li role="presentation" class="active"><a href="#members" aria-controls="members" role="tab" data-toggle="tab">{% trans 'Members' %}</a></li>
-        <li role="presentation"><a href="#moderators" aria-controls="moderators" role="tab" data-toggle="tab">{% trans 'Moderators' %}</a></li>
-    </ul>
-
-    <div class="tab-content">
-        <div role="tabpanel" class="tab-pane active" id="members">
-                <h2 class="sans-serif">{% trans 'Members' %}</h2>
-            <p>
-                <a href="{% url 'dashboard-project-invite' project.slug %}"
-                   class="btn btn-primary pull-right">{% trans 'invite users' %}</a>
-            </p>
-            {% if form.requests.forms or form.invites.forms or form.users.forms %}
-                {% csrf_token %}
-                <table class="table table-hover">
-                    <thead>
-                    <tr>
-                        <th>{% trans 'Name' %}</th>
-                        <th>{% trans 'Email' %}</th>
-                        <th>{% trans 'Edit' %}</th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    {% for form in form.requests.forms %}
-                        <tr>
-                            <td>
-                                <a href="{% url 'profile' form.instance.creator.username %}">
-                                    <img src="{% get_avatar form.instance.creator 'org_avatar_medium' %}" alt="" class="circled"> {{ form.instance.creator.username }}
-                                </a>
-                                <span class="label label-subtle">{% trans 'Request' %}</span>
-                            </td>
-                            <td>
-                                <strong>{{ form.instance.creator.email }}</strong>
-                            </td>
-                            <td>
-                                {{ form.id }}
-                                {% for radio in form.action %}
-                                    <div class="radio-action">
-                                        {{ radio.tag }}
-                                        <label for="{{ radio.id_for_label }}"
-                                               class="{{ radio.choice_label|lower }} radio-label">
-                                            <span class="radio-text sr-only">{{ radio.choice_label }}</span>
-                                        </label>
-                                    </div>
-                                {% endfor %}
-                            </td>
-                        </tr>
-                    {% endfor %}
-                    {% for form in form.invites.forms %}
-                        <tr>
-                            <td>
-                                <strong class="text-uppercase text-danger">{% trans 'Pending' %}</strong>
-                            </td>
-                            <td><strong>{{ form.instance.email }}</strong></td>
-                            <td>
-                                {{ form.id }}
-                                <div class="radio-action">
-                                    {{ form.delete }}
-                                    <label for="{{ form.delete.id_for_label }}"
-                                           class="{{ form.delete.label|lower }} radio-label">
-                                        <span class="radio-text sr-only">{{ form.delete.label }}</span>
-                                    </label>
-                                </div>
-                            </td>
-                        </tr>
-                    {% endfor %}
-                    {% for form in form.users.forms %}
-                        <tr>
-                            <td>
-                                <a href="{% url 'profile' form.instance.username %}">
-                                    <img src="{% get_avatar form.instance 'org_avatar_medium' %}" alt="" class="circled">
-                                    {{ form.instance.username }}
-                                </a>
-                            </td>
-                            <td>
-                                <strong>{{ form.instance.email }}</strong>
-                            </td>
-                            <td>
-                                {{ form.id }}
-                                <div class="radio-action">
-                                    {{ form.delete }}
-                                    <label for="{{ form.delete.id_for_label }}"
-                                           class="{{ form.delete.label|lower }} radio-label">
-                                        <span class="radio-text sr-only">{{ form.delete.label }}</span>
-                                    </label>
-                                </div>
-                            </td>
-                        </tr>
-                    {% endfor %}
-                    </tbody>
-                </table>
-                <input type="submit" class="btn btn-primary"
-                       value="{% trans 'submit' %}">
-            {% else %}
-                <p>{% trans 'There are no invites, requests or users in this project yet.' %}</p>
-            {% endif %}
-        </div>
-        <div role="tabpanel" class="tab-pane" id="moderators">
-            <h2 class="sans-serif">{% trans 'Moderators' %}</h2>
-{% endif %}
-            <div class="row">
-                <div class="col-sm-4">
-                    <div class="form-group">
-                        <label for="{{ form.moderators.user.id_for_label }}">{{ form.moderators.user.label }}</label>
-                        {{ form.moderators.user|add_class:'form-control' }}
-                        {{ form.moderators.user.errors }}
-                    </div>
-                </div>
-            </div>
-            {% react_user_list project.moderators.all project 'moderators' %}
-            {{ form.moderators.media }}
-{% if not project.is_public %}
-        </div>
-    </div>
-{% endif %}
-{{ form.requests.management_form }}
-{{ form.invites.management_form }}
-{{ form.users.management_form }}
-</form>
-{% endblock %}
