+++ euth_wagtail/settings/base.py
@@ -88,7 +88,7 @@ INSTALLED_APPS = [
     'euth.organisations.apps.OrganisationsConfig',
     'euth.projects.apps.ProjectsConfig',
     'euth.ideas.apps.IdeaConfig',
-    'euth.a4_dashboard.apps.A4DashboardConfig',
+    'euth.dashboard.apps.DashboardConfig',
     'euth.accounts.apps.AccountConfig',
     'euth.memberships.apps.MembershipsConfig',
     'euth.documents.apps.DocumentConfig',
